feat(agent): add built-in tool registry
